requirements.txt
streamlit==1.36.0
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.35.13
python-dotenv==1.0.1
httpx==0.27.2
tiktoken==0.7.0